Moved homepage logic to its own file
Moved homepage logic to its own file to prevent suspense error using useSearchParams
